Fixed: only send one event when a directory is created.
In the current implementation, the inotify-monitor sends 2 events when a directory is created. The notification contains one event with the "fsw_event_flag::IsDir" flag, the second event contains the "fsw_event_flag::Created" flag. The filenames in both notifications are different (one contains the parent directory, and the other one the created directory).